Improve textures for Map(); Replace int by long in s-hull-pro; Remove duplicated points during triangulation

+++ src/s_hull/s_hull_pro.h
@@ -24,15 +24,17 @@ email david@s-hull.org
 
 */
 
+// Global replace int->long by A.Balakin 7 August 2012 -- 64bit version can handle huge data arrays
+
 struct Triad
 {
-	int a,b, c;
-	int ab, bc, ac;  // adjacent edges index to neighbouring triangle.
+	long a,b, c;
+	long ab, bc, ac;  // adjacent edges index to neighbouring triangle.
 	float ro, R,C;
-	//std::set<int> idx;
+	//std::set<long> idx;
 	Triad() {a=b=c=ab=bc=ac=0;	ro=R=C=0;}	// added by A.Balakin 6 July 2012 -- uninitialised variable
-	Triad(int x, int y) : a(x), b(y),c(0), ab(-1), bc(-1), ac(-1), ro(-1), R(0), C(0) {}
-	Triad(int x, int y, int z) : a(x), b(y), c(z),  ab(-1), bc(-1), ac(-1), ro(-1), R(0), C(0) {}
+	Triad(long x, long y) : a(x), b(y),c(0), ab(-1), bc(-1), ac(-1), ro(-1), R(0), C(0) {}
+	Triad(long x, long y, long z) : a(x), b(y), c(z),  ab(-1), bc(-1), ac(-1), ro(-1), R(0), C(0) {}
 	Triad(const Triad &p) : a(p.a), b(p.b), c(p.c), ab(p.ab), bc(p.bc), ac(p.ac), ro(p.ro), R(p.R), C(p.C) {}
 
 	Triad &operator=(const Triad &p)
@@ -62,7 +64,7 @@ struct Triad
 
 struct Shx
 {
-	int id, trid;
+	long id, trid;
 	float r,c, tr,tc ;
 	float ro;
 	Shx() {r=c=tr=tc=ro=0;	id=trid=0;}	// added by A.Balakin 6 July 2012 -- uninitialised variable
@@ -98,12 +100,12 @@ inline bool operator<(const Shx &a, const Shx &b)
 
 struct Dupex
 {
-	int id;
+	long id;
 	float r,c;
 
 	Dupex() {}
 	Dupex(float a, float b) : id(-1), r(a), c(b) {}
-	Dupex(float a, float b, int x) : r(a), c(b), id(x) {}
+	Dupex(float a, float b, long x) : r(a), c(b), id(x) {}
 	Dupex(const Dupex &p) : id(p.id),  r(p.r), c(p.c) {}
 
 	Dupex &operator=(const Dupex &p)
@@ -132,18 +134,18 @@ inline bool operator<(const Dupex &a, const Dupex &b)
 // from s_hull.C
 
 
-int s_hull_pro( std::vector<Shx> &pts, std::vector<Triad> &triads);
+long s_hull_pro( std::vector<Shx> &pts, std::vector<Triad> &triads);
 void circle_cent2(float r1,float c1, float r2,float c2, float r3,float c3,float &r,float &c, float &ro2);
 void circle_cent4(float r1,float c1, float r2,float c2, float r3,float c3,float &r,float &c, float &ro2);
 void write_Shx(std::vector<Shx> &pts, char * fname);
 void write_Triads(std::vector<Triad> &ts, char * fname);
-int Cline_Renka_test(float &Ax, float &Ay, float &Bx, float &By, float &Cx, float &Cy, float &Dx, float &Dy);
-int T_flip_pro( std::vector<Shx> &pts, std::vector<Triad> &triads, std::vector<int> &slump, int numt, int start, std::set<int> &ids);
-int T_flip_pro_idx( std::vector<Shx> &pts, std::vector<Triad> &triads, std::vector<int> &slump, std::set<int> &ids);
+long Cline_Renka_test(float &Ax, float &Ay, float &Bx, float &By, float &Cx, float &Cy, float &Dx, float &Dy);
+long T_flip_pro( std::vector<Shx> &pts, std::vector<Triad> &triads, std::vector<long> &slump, long numt, long start, std::set<long> &ids);
+long T_flip_pro_idx( std::vector<Shx> &pts, std::vector<Triad> &triads, std::vector<long> &slump, std::set<long> &ids);
 
-int read_Shx(std::vector<Shx> &pts, char * fname);
+long read_Shx(std::vector<Shx> &pts, char * fname);
 
 
-int de_duplicate( std::vector<Shx> &pts,  std::vector<int> &outx );
+size_t de_duplicate( std::vector<Shx> &pts,  std::vector<size_t> &outx );
 
 #endif
